finished all packets for protocol 754

+++ src/fr/rader/bob/protocol/Packet.java
@@ -6,7 +6,7 @@ public interface Packet {
 
     int getLength();
 
-    byte getPacketID();
+    int getPacketID();
 
     int getTimestamp();
 }
